entities-interactor: rename ISerializedProgress
