+++ src/entities-interactor/ISerializedProgress.ts
